requirements.txt
requests~=2.32.3
ConfigArgParse~=1.7
PyYAML~=6.0.2
jsonschema~=4.23.0
GitPython~=3.1.44
pydantic~=2.10.6
